refactory account_parser
