+++ src/program-rust/Cargo.toml
@@ -17,10 +17,6 @@ arrayref = "0.3.6"
 num-derive = "0.3.3"
 num-traits = "0.2.14"
 num_enum = "0.5.6"
-#num = "0.4.0"
-safe-transmute = "0.11.0"
-bytemuck = { version = "1.7.3" }
-# enumflags2 = "0.7.3"
 lazy_static = "1.4.0"
