+++ src/program-rust/src/swappers/crema/instruction.rs
@@ -80,4 +80,3 @@ fn swap_instruction(
     data,
   })
 }
-
